README Separate out the instructions
Move the information about what this is, to a top section to give a better at-a-glance overview of what the repo is. Separate out the instructions, also by adding a new heading at the bottom "Errors and colors in this paint style" for that further info which is not instructions.